+++ modules/forum/lib/models/SectionRelations.php
@@ -1,0 +1,27 @@
+<?php
+
+/**
+ * This file is part of JohnCMS Content Management System.
+ *
+ * @copyright JohnCMS Community
+ * @license   https://opensource.org/licenses/GPL-3.0 GPL-3.0
+ * @link      https://johncms.com JohnCMS Project
+ */
+
+declare(strict_types=1);
+
+namespace Forum\Models;
+
+use Illuminate\Database\Eloquent\Relations\HasMany;
+
+trait SectionRelations
+{
+
+    /**
+     * Связь с подразделами
+     */
+    public function subsections(): HasMany
+    {
+        return $this->hasMany(self::class, 'parent', 'id');
+    }
+}